chore: move get_table_type in relation
