+++ tests/unit/test_relation.py
@@ -1,8 +1,28 @@
-from dbt.adapters.athena.relation import AthenaRelation
+import pytest
+
+from dbt.adapters.athena.relation import AthenaRelation, TableType, get_table_type
 
 from .constants import DATA_CATALOG_NAME, DATABASE_NAME
 
 TABLE_NAME = "test_table"
+
+
+class TestRelation:
+    def test__get_relation_type_table(self):
+        assert get_table_type({"Name": "name", "TableType": "table"}) == TableType.TABLE
+
+    def test__get_relation_type_with_no_type(self):
+        with pytest.raises(ValueError):
+            get_table_type({"Name": "name"})
+
+    def test__get_relation_type_view(self):
+        assert get_table_type({"Name": "name", "TableType": "VIRTUAL_VIEW"}) == TableType.VIEW
+
+    def test__get_relation_type_iceberg(self):
+        assert (
+            get_table_type({"Name": "name", "TableType": "EXTERNAL_TABLE", "Parameters": {"table_type": "ICEBERG"}})
+            == TableType.ICEBERG
+        )
 
 
 class TestAthenaRelation:
